bug #PSBM-89210 aka Dispatcher and snmpd huge memory consumption
extra data is stored along the request when a new record is being
inserted. we rely on correct snmp routine but current users don't
follow the rules and don't provide free request in the end. noboby
frees the memory and we have a leak. let's register a callback to
free the memory.