bug #PSBM-89621 aka "net-snmp" crash: "snmpd killed by SIGSEGV"
somebody called exit. it was not snmpd because it doesn't use
value 6 for its exit status code. thus plugins finalization has not
been performed properly. prlsdk was still running while some of
its dependencies have already been destroyed by glibc cleanup
handlers.
exit man suggests to register an atexit handler to perform cleanup
of a shared library which is the case for rmond. as these handlers
are executed in an order reverse to the one of their registration
we have a chance to not face semi-dismantled configuration and
deinit sdk gracefully.