bug #PSBM-92261 aka "net-snmp" crash: "snmpd killed by SIGSEGV"
we have a race between a find and a erase. if the erase wins it
may delete the row before the find increments its reference counter.
callee of the find will end with a bogus tuple.
seems it's time to protect storage accesses with a lock to prevent
these races