fixes a heap corruption (use after free)
netsnmp lib backs up a table row data in the request on traversal
of the table. this is ok in case of a sequential execution but
leads to ub if another thread removes the row from the table until
the data is being accessed.
the problem is we use a set to order table items thus it's impossible
to check for existence in the set.
let's have a simple hash for the purpose and make another test inside
the critical section to make sure the row is still in the table.