Added memory release functions for netsnmp_mib_handler #PSBM-139802
The netsnmp_mib_handler itself consists of a data pointer myvoid
and a function pointer data_free. data_free is used to tell the netsnmp
library to free data. We add a function to burn out the data,
some kind of wrapper for the case when calling the destructor is important.

Signed-off-by: Fima Shevrin <efim.shevrin@virtuozzo.com>